wikiminiatlas.js: Add license note and bump version
MediaWiki:Wikiminiatlas.js has been hosted on Meta-Wiki since 2007:
https://meta.wikimedia.org/wiki/MediaWiki:Wikiminiatlas.js?oldid=647988
and it was only copied into this Git repository in 2013 (1043e835c).

Thus presumably this part of WikiMiniAtlas is available under
the licenses used on Meta-Wiki, which are CC BY-SA 4.0 and GFDL,
and not the license used for the rest of the WikiMiniAtlas code.